ISSUE #2093 - Renamed folder issueReference -> ticketReference
