+++ frontend/routes/components/messagesList/components/message/components/markdownMessage/ticketReference/issueReference.styles.ts
